Upload a module to code sharing site and DockerHub

Look up usernames and other information contained in "om.yml" to upload module to a code sharing site (github, gitlab or bitbucket) and/or DockerHub.

module_upload(
  flpth = getwd(),
  code_sharing = TRUE,
  dockerhub = TRUE,
  verbose = TRUE
)

Arguments

flpth
File path to location of module

code_sharing
Upload to code sharing service?

dockerhub
Upload to DockerHub?

verbose
Print docker and program info to console

Value

Logical

Details

This function runs git_upload and docker_push.

See also

Examples

library(outsider)

# NOT RUN
# # build a simple module
# module_path <- module_skeleton(program_name = 'echo', flpth = getwd(),
#                                repo_user = '[YOUR USERNAME]',
#                                docker_user = '[YOUR USERNAME]',
#                                service = 'github')
# module_build(module_path)
# # upload to your GitHub and your DockerHub
# module_upload(flpth = module_path)