Upload module to code-sharing service

Upload module to a git-based code-sharing service. Initiate a git repo, add core module files, commit and push to remote.

git_upload(flpth, username, service = c("github", "gitlab", "bitbucket"))

Arguments

flpth
File path to module.

username
Username for code-sharing service.

service
Code-sharing service

Value

Logical

Details

Remote URL is determined to be: code sharing URL + username + R package name. Git must be configured on a user's system before this function can be run.

See also

Examples

# Git must be configured before function can be run.
# To configure, provide your name and email via the command-line, e.g.
# git config --global user.name "John Doe"
# git config --global user.email johndoe@example.com
# To check your username and email, try:
# git config --list

# NOT RUN
# construct a simple module
module_path <- module_skeleton(program_name = 'echo', flpth = getwd())
module_check(flpth = module_path)

#> DESCRIPTION found ✓
#> R folder with files found ✓
#> inst found ✓
#> inst/om.yml found ✓
#> inst/dockerfiles found ✓
#> inst/dockerfiles/latest with one Dockerfile found ✓

module_identities(flpth = module_path)

#> R package name: 'om..echo'
#> Docker images: '/om_echo:latest'

# after these steps all files are built, to upload to git use:
git_upload(flpth = module_path, username = '[YOUR USERNAME]',
           service = '[SERVICE]')

#> Error in match.arg(service): 'arg' should be one of “github”, “gitlab”, “bitbucket”